services/blog/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "Blog service with gRPC interface"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.51.0"
grpcio-tools = "^1.51.0"
pydantic = "^2.0.0"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
